Interim commit - first pass of Moveable and Renderable Systems and Components

+++ VikingMathApp/VMAHealthSystem.m
@@ -10,7 +10,7 @@
 
 @implementation VMAHealthSystem
 
-- (void)update:(float)dt
+- (void)update:(double)dt
 {
 
 }
